requirements.txt
chardet
pyyaml
pytest
pytest-cov